2020-2021/StudentProjects/Echipa07/application/backend/requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
click==7.1.2
cycler==0.10.0
dnspython==1.16.0
eventlet==0.29.1
Flask==1.1.2
Flask-SocketIO==4.3.2
gast==0.2.2
google-pasta==0.2.0
greenlet==0.4.17
grpcio==1.34.0
h5py==2.10.0
imageai==2.1.5
importlib-metadata==3.3.0
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.18.5
opencv-python==4.4.0.46
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==3.14.2
python-socketio==4.6.1
PyYAML==5.3.1
scipy==1.5.4
six==1.15.0
tensorboard==1.15.0
tensorflow==1.15.4
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.4
termcolor==1.1.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
